refactor: move cli scripts into package
